Create readTable function inside table.go
Reading the `sqlite_schema` table is actually no
different to how we read any other table, and
doing this allows us to read other tables as well
provided we have the root page for them.

I've also introduced a new idea of 'raw records'
which basically is the entire byte slice of a
record, rather than just the bytes which are
defined directly within a cell. I think this is
more what I wanted to do when I originally wrote
the `cell.go` file. It allows us to detach the
idea of pages / cells from the idea of records,
which I think is good!