Separate layers in different packages
We are creating several different layers that will
each handle different parts of processing a
command:

- The parser takes an SQL string an returns an AST
- The generator takes an AST and returns an
  exeuction plan which is a list of instructions
- The machine executes instructions
- The btree layer handles reading and writing from
  btrees (tables or indexes)
- The pager handles reading/writing/caching pages
  from the disk

I'm not completely convinced that this will be the
final architecture, but it seems good for now and
pretty closely resembles the architecture of
SQLite.

I'd quite like to look into the idea of creating
public interfaces for each of these layers, so
that there are no dependencies between them.
Ideally, a (non-public) change in one package
should only cause that package to need to be
recompiled.