Define Pager and BTreeEngine interfaces in caller packages
This allows for a complete separation of these
packages! The BTree package does not import the
pager package anywhere, and the Machine package
does not import the BTree package.

They are completely separate layers that are
designed to conform to the interface defined by
the layers above them. In other words, a layer
understands how to use the layer below it (by
using the functions defined in the interface), but
doesn't know or care how the layer below it
implements those functions.

I've also created a new types package that has an
Entry type. This is how we represent an 'Entry' in
a database file (which doesn't have a single type,
it can be a string / integer / etc). My rationale
for making this a top level package is that it's
something that could be used by multiple layers in
the same way that a `int` or `string` could be.
It's almost just another primitive type!